circle buffer for snake body
i wanted to reduce the number of times i have to change the x,y pos of the snake body each time the snake moves.

in theory, i should only care about the head and tail. when the head advances, remove the current tail.

i tried this approach on the game board and it works well. however this doesnt work a fixed array because the *head* doesnt actually move in the array, the x,y vars for the head change instead.

so i decided to actually move the head around in the array, wrapping around using mod(array_size) to wrap around.

while this is overkill for something like snake, this still scales really well if i decided to make the snake body 1million in size. i wouldn't have to shift each element, only the head ;)