Fix direction swap bug
because the body of the snake just follows the head, it wouldn't make sense for each body entry to keep track of the direction.

however, if i want to reverse the snake direction, then i need to know the direction of the tail and invert it.

and because the tail direction is basically buffered by the size of the body, i have to keep track of the direction in each body part :(